requirements.txt
googleapis-common-protos==1.62.0
grpclib==0.4.6
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
multidict==6.0.4
paho-mqtt==1.6.1
Pillow==10.1.0
protobuf==4.25.1
typing_extensions==4.9.0
viam-sdk==0.12.0
